refactor(traces): the last two traces-only hooks come home (LFE-14804)
src/hooks/{useParsedObservation,usePreserveRelativeScroll}.ts are used by
nothing but traces, so the shared hooks folder was claiming them for no reason.
rule 6 in traces: 3 -> 1.

The one that remains is src/components/table/use-cases/scores.tsx, which needs a
judgment call rather than a move: it is a table use-case by shape and traces-only
by usage.
