+++ web/src/features/traces/components/IOPreview/IOPreview.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { type ScoreDomain, type Prisma } from "@langfuse/shared";
 import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePostHogClientCapture";
 import useLocalStorage from "@/src/components/useLocalStorage";
-import { usePreserveRelativeScroll } from "@/src/hooks/usePreserveRelativeScroll";
+import { usePreserveRelativeScroll } from "@/src/features/traces/hooks/usePreserveRelativeScroll";
 import { type MediaReturnType } from "@/src/features/media/validation";
 import { type ExpansionState } from "@/src/features/traces/components/AdvancedJsonViewer/types";
