+++ web/src/features/traces/hooks/useParsedObservation.ts
